+++ App.js
@@ -1,103 +1,29 @@
-import { React, useState, useEffect } from "react";
-import { StyleSheet, Text, View, Alert, TouchableOpacity } from "react-native";
-import { NavigationContainer } from "@react-navigation/native";
-import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
-import { createStackNavigator } from "@react-navigation/stack";
-import Login from "./screens/Login";
-import Record from "./screens/Record";
-import Centre from "./screens/Centre";
-import SignOut from "./screens/SignOut";
-import firebase from "firebase";
-import { Ionicons } from 'react-native-vector-icons';
-import { Platform } from 'react-native';
-import AsyncStorage from '@react-native-async-storage/async-storage';
+import { StatusBar } from 'expo-status-bar';
+import React from 'react';
+import { StyleSheet, Text, View } from 'react-native';
+import { NavigationContainer } from '@react-navigation/native';
+import { createNativeStackNavigator } from '@react-navigation/native-stack';
+import LoginScreen from './screens/LoginScreen';
+import HomeScreen from './screens/HomeScreen';
 
-
-const Tab = createBottomTabNavigator();
-const Stack = createStackNavigator();
+const Stack = createNativeStackNavigator();
 
 export default function App() {
-  const [loggedIn, setLoggedIn] = useState(false);
-
-  useEffect(() => {
-    const unsubscribe = firebase.auth().onAuthStateChanged((user) => {
-      setLoggedIn(!!user);
-    });
-    return unsubscribe;
-  }, []);
-
-  function AuthStack() {
-    return (
+  return (
+    <NavigationContainer>
       <Stack.Navigator>
-        <Stack.Screen name="Login" component={Login} />
-        <Stack.Screen name="Record" component={Record} />
+        <Stack.Screen options={{ headerShown: false }} name="Login" component={LoginScreen} />
+        <Stack.Screen name="Home" component={HomeScreen} />
       </Stack.Navigator>
-    );
-  }
-
-  if (!loggedIn) {
-    return (
-      <NavigationContainer>
-        <AuthStack />
-      </NavigationContainer>
-    );
-  }
-
-  if (Platform.OS === 'ios' || 'android') {
-    return (
-      <NavigationContainer>
-        <Tab.Navigator
-          screenOptions={({ route }) => ({
-            headerShown: false,
-            tabBarIcon: ({ focused, color, size }) => {
-              let iconName;
-              if (route.name === 'Record') {
-                iconName = focused
-                  ? 'mic-outline'
-                  : 'mic';
-              } else if (route.name === 'Centre') {
-                iconName = focused ? 'cloud' : 'cloud-outline';
-              } else if (route.name === 'Sign Out') {
-                iconName = focused ? 'exit' : 'exit-outline';
-              }
-              // You can return any component that you like here!
-              return <Ionicons name={iconName} size={size} color={color} />;
-            },
-          })}>
-          <Tab.Screen name="Record" component={Record} />
-          <Tab.Screen name="Centre" component={Centre} />
-          <Tab.Screen name="Sign Out" component={SignOut} />
-        </Tab.Navigator>
-      </NavigationContainer>
-    );
-
-  } else {
-    return (
-      <NavigationContainer>
-        <Tab.Navigator
-          screenOptions={({ route }) => ({
-            headerShown: false,
-            tabBarIcon: ({ focused, color, size }) => {
-              let iconName;
-              if (route.name === 'Centre') {
-                iconName = focused ? 'cloud' : 'cloud-outline';
-              }
-              // You can return any component that you like here!
-              return <Ionicons name={iconName} size={size} color={color} />;
-            },
-          })}>
-          <Tab.Screen name="Centre" component={Centre} />
-        </Tab.Navigator>
-      </NavigationContainer>
-    );
-  }
+    </NavigationContainer>
+  );
 }
 
 const styles = StyleSheet.create({
   container: {
     flex: 1,
-    backgroundColor: "#fff",
-    alignItems: "center",
-    justifyContent: "center",
+    backgroundColor: '#fff',
+    alignItems: 'center',
+    justifyContent: 'center',
   },
 });
