--- cors.json
@@ -1,5 +1,0 @@
-[    {        "origin": ["*"],
-        "method": ["GET"],
-        "maxAgeSeconds": 3600
-    }
-]
