--- firebase/auth.js
@@ -1,6 +1,0 @@
-import "firebase/auth";
-import firebase from "./firebase";
-import firebaseConfig from "./firebaseConfig";
-
-
-export default firebase.auth();
