--- firebase/firebase.js
@@ -1,6 +1,0 @@
-import firebase from 'firebase/app';
-import firebaseConfig from './firebaseConfig';
-
-const app = firebase.initializeApp(firebaseConfig);
-
-export default firebase;
