--- firebase/firebaseConfig.js
@@ -1,12 +1,0 @@
-const firebaseConfig = {
-    apiKey: "AIzaSyBajv536FPDjRNzqwZLNE0Pd_C-kC100z0",
-    authDomain: "ideacentre-97c4b.firebaseapp.com",
-    databaseURL: "https://ideacentre-97c4b-default-rtdb.firebaseio.com",
-    projectId: "ideacentre-97c4b",
-    storageBucket: "ideacentre-97c4b.appspot.com",
-    messagingSenderId: "825853901869",
-    appId: "1:825853901869:web:bc990ed11c36c052fbcca4",
-    measurementId: "G-8LCZWHB73M"
-};
-
-export default firebaseConfig;
