+++ package.json
@@ -9,32 +9,10 @@
     "web": "expo start --web"
   },
   "dependencies": {
-    "@expo/vector-icons": "^13.0.0",
-    "@expo/webpack-config": "^0.17.2",
-    "@firebase/firestore": "^3.8.4",
-    "@react-native-async-storage/async-storage": "^1.18.1",
-    "@react-navigation/bottom-tabs": "^6.5.7",
-    "@react-navigation/native": "^6.1.6",
-    "@react-navigation/native-stack": "^6.9.9",
-    "@react-navigation/stack": "^6.3.12",
-    "expo": "~47",
-    "expo-av": "~13.0.3",
-    "expo-file-system": "~15.1.1",
-    "expo-sharing": "~11.0.1",
+    "expo": "~47.0.8",
     "expo-status-bar": "~1.4.2",
-    "firebase": "^8.2.3",
     "react": "18.1.0",
-    "react-dom": "18.1.0",
-    "react-native": "0.70.5",
-    "react-native-safe-area-context": "4.4.1",
-    "react-native-screens": "~3.18.0",
-    "react-native-sound": "^0.11.2",
-    "react-native-table-component": "^1.2.2",
-    "react-native-vector-icons": "^9.2.0",
-    "react-native-web": "~0.18.9",
-    "react-navigation": "^4.4.4",
-    "react-navigation-stack": "^2.10.4",
-    "rn-fetch-blob": "^0.12.0"
+    "react-native": "0.70.5"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9"
